+++ src/main/java/com/codeclan/example/CoffeeTracker/repositories/CoffeeRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CoffeeRepository extends JpaRepository<Coffee, Long> {
-    List<Coffee> findCoffeesByStrength(Long strength);
+    List<Coffee> findCoffeesByStrength(Integer strength);
 }
